Make EditMemoDialog handle when the confirm button is disabled.
The EditMemoDialog now disables the Confirm button if one of the
following conditions is met:
 a) The title is not in use.
 b) The title has changed from its old version.
 c) The description has changed.
 d) The list of tags has been modified.

If using the dialog to create a new Memo only the first condition is
checked. When editing a new Memo all conditions are checked against. In
order to be able to compare new values to old values the old values are
now kept cached during the dialog's runtime.