Add a simple color picker and "add" button to ManageTagsDialog's UI.
The color picker is not wired to any logic related to colors yet. The
"add" button and the color picker are disabled if the search bar is
empty. The "add" button changes text based on what's in the searchbar or
what tag in the list is selected. If the text in the search bar does not
match any existing tags, the button will read "Add", if the text
matches an existing tag it will read "Update".
The text from the search bar is also displayed in a label.